-Moved ListItem component to shoppingList -Moved and fixed delete method
#6rdveu[In Progress]
